+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features
 
 - Added support for the `NO_COLOR` environment variable, see #1021 and #1031 (@eth-p)
+- Added `-P` short flag to disable paging, revised man page description, see #1075 and #1082 (@LordFlashmeow) 
 
 ## Bugfixes